# add define.exports