--- a/README.md
+++ b/README.md
@@ -17,6 +17,77 @@ loader.noConflict({
 });
 ```
 
+## Extra stuff
+
+### `define.alias('new-name', 'old/path')`
+
+`define.alias` allows creation of a symlink from one module to another, for example:
+
+```js
+define('foo/bar/baz', [], () => 'hi');
+define.alias('foo', 'foo/bar/baz');
+
+require('foo') // => 'hi';
+require('foo') === require('foo/bar/baz');
+```
+
+### `define.exports('foo', {})`
+
+`define.exports` enables a fastpath for non-lazy dependency-less modules, for example:
+
+Rather then:
+
+```js
+define("my-foo-app/templates/application", ["exports"], function (exports) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+
+  exports.default = Ember.HTMLBars.template({ "id": "VVZNWoRm", "block": "{\"statements\":[[1,[26,[\"welcome-page\"]],false],[0,\"\\n\"],[0,\"\\n\"],[1,[26,[\"outlet\"]],false]],\"locals\":[],\"named\":[],\"yields\":[],\"hasPartials\":false}", "meta": { "moduleName": "my-foo-app/templates/application.hbs" } });
+});
+```
+
+We can author:
+
+```js
+define.exports('my-app/template/apple', { hbs: true, "id": "VVZNWoRm", "block": "{\"statements\":[[1,[26,[\"welcome-page\"]],false],[0,\"\\n\"],[0,\"\\n\"],[1,[26,[\"outlet\"]],false]],\"locals\":[],\"named\":[],\"yields\":[],\"hasPartials\":false}", "meta": { "moduleName": "my-foo-app/templates/application.hbs" }});
+```
+
+benefits:
+
+* less bytes
+* no reification step
+* no need to juggle pre-parse voodoo.
+
+### `require.unsee('foo');
+
+`require.unsee` allows one to unload a given module. *note* The side-affects of that module cannot be unloaded.
+This is quite useful, especially for test suites. Being able to unload run tests, mitigates many common memory leaks:
+
+example:
+
+```js
+define('my-app/tests/foo-test.js', ['qunit'], function(qunit) {
+  let app;
+
+  qunit.module('my module', {
+    beforeEach() {
+      app = new App();
+    }
+
+    // forgot to `null` out app in the afterEach
+  });
+
+  test('my app exists', function(assert) {
+    assert.ok(app);
+  })
+})
+```
+
+---
+
 Note: To be able to take advantage of alternate `define` method name, you will also need to ensure that your
 build tooling generates using the alternate.  An example of this is done in the [emberjs-build](https://github.com/emberjs/emberjs-build)
 project in the [babel-enifed-module-formatter plugin](https://github.com/emberjs/emberjs-build/blob/v0.4.2/lib/utils/babel-enifed-module-formatter.js).