--- a/lib/loader/loader.js
+++ b/lib/loader/loader.js
@@ -245,6 +245,26 @@ var loader, define, requireModule, require, requirejs;
     heimdall.stop(token);
   };
 
+  define.exports = function(name, defaultExport) {
+    var module = registry[name];
+
+    // If a module for this name has already been defined and is in any state
+    // other than `new` (meaning it has been or is currently being required),
+    // then we return early to avoid redefinition.
+    if (module && module.state !== 'new') {
+      return;
+    }
+
+    module = new Module(name, [], noop, null);
+    module.module.exports = defaultExport;
+    module.state = 'finalized';
+    registry[name] = module;
+
+    return module;
+
+  };
+
+  function noop() { }
   // we don't support all of AMD
   // define.amd = {};
 
@@ -336,7 +356,9 @@ var loader, define, requireModule, require, requirejs;
   define.alias('foo', 'foo/qux');
   define('foo/bar',  ['foo', './quz', './baz', './asdf', './bar', '../foo'], function() {});
   define('foo/main', ['foo/bar'], function() {});
+  define.exports('foo/exports', {});
 
+  require('foo/exports');
   require('foo/main');
   require.unsee('foo/bar');