@rwjblue 2017-05-25
Where does `Ember.HTMLBars.template(` bit go in the new semantics?