@stefanpenner 2017-05-25
> Where does Ember.HTMLBars.template( bit go in the new semantics?

I would move it to the caller, which would detect `require(templatePath).default.hbs === true` or something.