@stefanpenner 2017-05-25
Some numbers (note the benchmarks only care about runtime cost)

```sh
Benchmark
  file: ./benchmarks/scenarios/define-many.js
  total:  73ms
  per op:  7.3ms

Benchmark
  file: ./benchmarks/scenarios/define-many-exports.js
  total:  47ms
  per op:  4.7ms
```

---

File size (post gzip)

```sh
 t/size cat before.js | gzip -9 | wc -c
     318

 t/size cat after.js | gzip -9 | wc -c
     223
```


```js
// before.js
define("my-foo-app/templates/application", ["exports"], function (exports) {
  "use strict";

  Object.defineProperty(exports, "__esModule", {
    value: true
  });

  exports.default = Ember.HTMLBars.template({ "id": "VVZNWoRm", "block": "{\"statements\":[[1,[26,[\"welcome-page\"]],false],[0,\"\\n\"],[0,\"\\n\"],[1,[26,[\"outlet\"]],false]],\"locals\":[],\"named\":[],\"yields\":[],\"hasPartials\":false}", "meta": { "moduleName": "my-foo-app/templates/application.hbs" } });
});
```

```js
// after.js
define.exports('my-app/templates/my-template.hbs', { hbs: true, "id": "VVZNWoRm", "block": "{\"statements\":[[1,[26,[\"welcome-page\"]],false],[0,\"\\n\"],[0,\"\\n\"],[1,[26,[\"outlet\"]],false]],\"locals\":[],\"named\":[],\"yields\":[],\"hasPartials\":false}", "meta": { "moduleName": "my-foo-app/templates/application.hbs" }});
```