@rwjblue 2017-05-25
> I would move it to the caller, which would detect require(templatePath).default.hbs === true or something.

Maybe, but if we do this, then it needs to include version info or something. Should be easy to put into the meta though...