@stefanpenner 2017-05-25
> Maybe, but if we do this, then it needs to include version info or something. Should be easy to put into the meta though...

totally agree.  I'll run more thorough tests to confirm/deny my intuition (before we proceed with the required plumbing)