@rwjblue 2017-05-25
Another issue that we will have to address is how the templates get instantiated. In older Ember versions, templates were not instantiated but as of Ember 2.10 they are instantiated (and receive injections). Today, `Ember.HTMLBars.template` returns an object that adheres to the standard DI system contract (and we create instances of the templates by calling `.create(injectionsHere)` on them).