@stefanpenner 2017-05-25
> Today, Ember.HTMLBars.template returns an object that adheres to the standard DI system contract (and we create instances of the templates by calling .create(injectionsHere) on them).

Ya, I think we can wrap on the caller side for this as well.