@stefanpenner 2017-06-15
released as v4.5.0 🎉